requirements.txt
# Production dependencies
Flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0

# Database - SQLAlchemy ORM
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP requests (for Mistral API)
requests==2.31.0
